Fix possibility of orphaned derivative files in S3.
For newly created photo records, the manual derivative process was
leading to the files being uploaded without an ID first, and then moved
to the ID based path. However, this led to orphaned derivative files
sticking around after the save.

In order to fix this, rely on the derivative create on promotion
process. This wasn't working for us originally, since it happens in an
`after_commit` callback. This means our progress was being reported as
done too soon (since it occurred inside the transaction but before the
derivatives could be created). So fix all this by updating the progress
outside of the transaction so it can occur after the automatic
derivative process.